Refactoring
- Move Request, Response, Status, HTTPError to top uzhttp package
- Move websocket Frame into its own package
- Move WebsocketResponse into Response
- Split some things into their own source file
